Moved forms into viewmodels package
