+++ app/viewmodels/LoginData.java
@@ -1,9 +1,9 @@
-package controllers.forms;
+package viewmodels;
 
 import play.data.validation.Constraints;
 import services.ConsumerService;
 
-public class Login {
+public class LoginData {
 
     @Constraints.Required(message="Bitte gib eine Emailadresse an.")
     @Constraints.Email(message = "Bitte gib eine gültige Emailadresse an.")
