+++ app/views/ConsumerManagement/login.scala.html
@@ -1,4 +1,4 @@
-@(loginForm: Form[controllers.forms.Login])
+@(loginForm: Form[viewmodels.LoginData])
 @import helper._
 
 @main("BringToAfrica") {
