+++ app/views/ConsumerManagement/registration.scala.html
@@ -1,4 +1,4 @@
-@(registerForm: Form[controllers.forms.Registration])
+@(registerForm: Form[viewmodels.RegistrationData])
 @import helper._
 
 @main("BringToAfrica") {
