+++ app/views/newProject.scala.html
@@ -1,4 +1,4 @@
-@(projektDataForm: Form[controllers.forms.ProjectData])
+@(projektDataForm: Form[viewmodels.ProjectData])
 @import helper._
 
 @main("BringToAfrica") {
