+++ test/Integration/NewProjectTest.java
@@ -1,22 +1,12 @@
 package Integration;
 
 
-import models.Project;
 import models.Consumer;
 import org.junit.Test;
 import org.openqa.selenium.By;
-import org.openqa.selenium.Cookie;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.support.ui.WebDriverWait;
-
-import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-
-import static play.test.Helpers.*;
-
 
 public class NewProjectTest {
     @Test
